Video: Fix warnings in ImageMagick

+++ src/platform/imagemagick/imagemagick-gif-encoder.c
@@ -8,7 +8,7 @@
 #include "gba/video.h"
 
 static void _magickPostVideoFrame(struct GBAAVStream*, struct GBAVideoRenderer* renderer);
-static void _magickPostAudioFrame(struct GBAAVStream*, int32_t left, int32_t right);
+static void _magickPostAudioFrame(struct GBAAVStream*, int16_t left, int16_t right);
 
 void ImageMagickGIFEncoderInit(struct ImageMagickGIFEncoder* encoder) {
 	encoder->wand = 0;
@@ -70,7 +70,7 @@ static void _magickPostVideoFrame(struct GBAAVStream* stream, struct GBAVideoRen
 	++encoder->currentFrame;
 }
 
-static void _magickPostAudioFrame(struct GBAAVStream* stream, int32_t left, int32_t right) {
+static void _magickPostAudioFrame(struct GBAAVStream* stream, int16_t left, int16_t right) {
 	UNUSED(stream);
 	UNUSED(left);
 	UNUSED(right);
